+++ render/intersections/inter_sphere.c
@@ -2,9 +2,9 @@
 
 void 	inter_sphere(t_sphere *s_sphere, t_ray *s_ray)
 {
-	double	b;
-	double	c;
-	double	d_x;
+	float	b;
+	float	c;
+	float	d_x;
 	t_ray	s_ray_inter;
 
 	s_ray_inter.s_vec_inter = vec_sub(&s_ray->s_vec_start, s_sphere->s_vec_origin);
@@ -14,7 +14,7 @@ void 	inter_sphere(t_sphere *s_sphere, t_ray *s_ray)
 	d_x = b * b - (4 * c);
 	if (d_x < 0)
 		return ;
-	d_x = (-b - sqrt(d_x)) * 0.5;
+	d_x = (float)((-b - sqrt((double)d_x)) * 0.5);
 	if (d_x >= 0.0001 && d_x < s_ray->length)
 	{
 		s_ray_inter = *s_ray;
